refactor(app): changed base project's name to My Scent App
