+++ app/src/main/java/com/hika/myscentapp/MyScent.kt
@@ -1,4 +1,4 @@
-package com.hika.myscent
+package com.hika.myscentapp
 
 import android.app.Application
 import com.hika.admin.di.adminModule
@@ -6,7 +6,7 @@ import com.hika.auth.di.authModule
 import com.hika.data.di.databaseModule
 import com.hika.data.di.firebaseModule
 import com.hika.data.di.repositoryModule
-import com.hika.myscent.di.navigatorModule
+import com.hika.myscentapp.di.navigatorModule
 import com.hika.user.di.userModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
