+++ app/src/main/java/com/hika/myscentapp/di/NavigatorModule.kt
@@ -1,10 +1,10 @@
-package com.hika.myscent.di
+package com.hika.myscentapp.di
 
 import com.hika.admin.navigation.AdminNavigation
 import com.hika.auth.navigation.AuthNavigation
-import com.hika.myscent.navigator.AdminNavigator
-import com.hika.myscent.navigator.AuthNavigator
-import com.hika.myscent.navigator.UserNavigator
+import com.hika.myscentapp.navigator.AdminNavigator
+import com.hika.myscentapp.navigator.AuthNavigator
+import com.hika.myscentapp.navigator.UserNavigator
 import com.hika.user.navigation.UserNavigation
 import org.koin.dsl.module
