+++ app/src/main/java/com/hika/myscentapp/navigator/AuthNavigator.kt
@@ -1,4 +1,4 @@
-package com.hika.myscent.navigator
+package com.hika.myscentapp.navigator
 
 import com.hika.auth.navigation.AuthNavigation
