+++ app/src/main/java/com/hika/myscentapp/navigator/UserNavigator.kt
@@ -1,12 +1,13 @@
-package com.hika.myscent.navigator
+package com.hika.myscentapp.navigator
 
 import android.app.Activity
-import com.hika.admin.navigation.AdminNavigation
 import com.hika.auth.navigation.AuthNavigation
+import com.hika.user.navigation.UserNavigation
 
-class AdminNavigator(
+class UserNavigator(
     private val authNavigation: AuthNavigation
-): AdminNavigation() {
+): UserNavigation() {
+
     override fun navigateToAuth(activity: Activity) {
         authNavigation.navigateToAuth(activity)
     }
